refactor(geometry): split processors/boolean.rs into core + carve-outs
boolean.rs (1741 LOC) split into a boolean/ directory: the two self-contained
concerns move out - halfspace_cap.rs (cap_half_space_clip, 267) with its tests
in halfspace_cap_tests.rs (223), and polygonal_prism.rs (the PBHS prism
builder, 378). The recursive CSG processor core stays in mod.rs (894): its
process_with_depth dispatch + chained-clip union path + failure plumbing are
tightly coupled through the per-instance self.failures RefCell, so it is one
cohesive unit (splitting it would spread shared mutable state across files).

Pure code motion (only super::/import/visibility scaffolding changed); one
pub(super) bump (build_polygonal_bounded_half_space_mesh). cargo build/clippy
-D warnings clean; targeted boolean/cap tests green.
